Extract a big nested class, fix spelling errors.
